pyproject.toml
[project]
name = "bestpractice"
version = "0.1.0"
description = ""
authors = [
    {name = "Yasmine Atrous",email = "yasmine.atrous@esprit.tn"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1)",
    "pypdf2 (>=3.0.1)",
    "python-docx (>=1.2.0)",
    "uvicorn (>=0.35.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "llama-parse (>=0.6.46,<0.7.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
